backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
google-genai
python-multipart
gTTS
sqlalchemy
